+++ packages/seahorse/src/primitives/index.ts
@@ -27,6 +27,9 @@ export * from './select'
 export * from './form-control'
 export * from './spinner'
 
+// Lists
+export * from './flash-list'
+
 // Feedback & display
 export * from './modal'
 export * from './alert-dialog'
@@ -37,3 +40,4 @@ export * from './skeleton'
 export * from './card'
 export * from './notification-dot'
 export * from './toast'
+export * from './image-background'